backend/requirements.txt
flask[async]
flask-cors
boto3
asyncpg
python-dotenv
sagemaker
pinecone
gunicorn 
openai
numpy